third commit

+++ src/BmiService.java
@@ -1,10 +1,10 @@
 public class BmiService {
 
     public double calculate(double meters, double kg) {
-        int bodyIndex;
-        bodyIndex = (int) (kg / (meters * meters));
+        double bodyIndex;
+        bodyIndex =  (kg / (meters * meters));
 
-        return bodyIndex;
+        return  (int) bodyIndex;
     }
 }
